build(scm-diff-editor): move most symbols to `lib.rs`
- Many of these symbols were duplicated by accident.
- I forgot that git-branchless links directly against `scm-diff-editor`'s code, not just `scm-record`'s code.
